Reorg asset files
