+++ app/assets/stylesheets/main.scss
@@ -1,2 +1,3 @@
 // Partials
 @import "partials/developers.scss";
+@import "partials/datatables.scss";
